requirements.txt
asgiref==3.7.2
attrs==23.1.0
build==0.10.0
CacheControl==0.12.14
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.4.1
distlib==0.3.6
Django==4.2.2
django-celery-beat==2.5.0
django-cors-headers==4.1.0
django-filter==23.3
django-phonenumber-field==7.1.0
django-split-settings==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
dulwich==0.21.5
filelock==3.12.2
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==6.7.0
inflection==0.5.1
installer==0.7.0
itypes==1.2.0
jaraco.classes==3.2.3
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
MarkupSafe==2.1.3
more-itertools==9.1.0
msgpack==1.0.5
packaging==23.1
pexpect==4.8.0
phonenumbers==8.13.14
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.6.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.7.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
rapidfuzz==2.15.1
redis==4.6.0
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
shellingham==1.5.0.post1
six==1.16.0
sqlparse==0.4.4
tldextract==3.4.4
tomli==2.0.1
tomlkit==0.11.8
trove-classifiers==2023.5.24
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.23.1
webencodings==0.5.1
xattr==0.10.1
zipp==3.15.0
zope.event==4.6
zope.interface==6.0

